Add files via upload
Design and implement C++/Java Program to find a subset of 
a given set S = {Sl, S2,…, Sn} of n positive integers whose 
SUM is equal to a given positive integer d. For example, if S 
= {1, 2, 5, 6, 8} and d= 9, there are two solutions {1, 2, 6} 
and {1, 8}. Display a suitable message, if the given problem 
instance doesn't have a solution.